Оптимизация форм. Перенос заказов в отдельную форму
