+++ MyStore/MyStore/OrdersListWindow.xaml
@@ -1,27 +1,24 @@
-﻿<Window x:Class="MyStore.EmployeesListWindow"
+﻿<Window x:Class="MyStore.OrdersListWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:MyStore"
         mc:Ignorable="d"
-        Title="Сотрудники" Height="450" Width="600" WindowStartupLocation="CenterScreen">
-        <Window.Resources>
-            <Style TargetType="Button">
-                <Setter Property="Margin" Value="5 5 0 2" />
-            </Style>
+        Title="Заказы" Height="300" Width="425" WindowStartupLocation="CenterScreen">
+    <Window.Resources>
+        <Style TargetType="Button">
+            <Setter Property="Margin" Value="5 5 0 2" />
+        </Style>
     </Window.Resources>
     <Grid>
         <DockPanel>
-            <ListView x:Name="employeesList" HorizontalAlignment="Stretch" DockPanel.Dock="Top" ItemsSource="{Binding}">
+            <ListView x:Name="ordersList" HorizontalAlignment="Stretch" DockPanel.Dock="Top" ItemsSource="{Binding}">
                 <ListView.View>
                     <GridView>
-                        <GridViewColumn DisplayMemberBinding="{Binding LastName}">Фамилия</GridViewColumn>
-                        <GridViewColumn DisplayMemberBinding="{Binding FirstName}">Имя</GridViewColumn>
-                        <GridViewColumn DisplayMemberBinding="{Binding MiddleName}">Отчество</GridViewColumn>
-                        <GridViewColumn DisplayMemberBinding="{Binding BirthDate}">Дата рождения</GridViewColumn>
-                        <GridViewColumn DisplayMemberBinding="{Binding Sex}">Пол</GridViewColumn>
-                        <GridViewColumn DisplayMemberBinding="{Binding Departament.Name}">Подразделение</GridViewColumn>
+                        <GridViewColumn DisplayMemberBinding="{Binding Number}">Номер</GridViewColumn>
+                        <GridViewColumn DisplayMemberBinding="{Binding ProductName}">Название товара</GridViewColumn>
+                        <GridViewColumn DisplayMemberBinding="{Binding EmployeeToString}">Сотрудник</GridViewColumn>
                     </GridView>
                 </ListView.View>
             </ListView>
